Fixed a couple of bugs with testing. Added a skills config file. Will continue to work on this. Our system is heavily simplified and we need to make sure it follows pathfinder rules.
