+++ pathfinder_sim/action_types.py
@@ -1,0 +1,9 @@
+# action_types.py
+from enum import Enum
+
+class ActionType(Enum):
+    STANDARD = "standard"
+    MOVE = "move"
+    SWIFT = "swift"
+    FREE = "free"
+    FULL_ROUND = "full_round"
