+++ pathfinder_sim/character.py
@@ -99,17 +99,7 @@ class Character:
         return [cond.get_status() for cond in self.conditions]
 
     def recalc_stats(self) -> None:
-        """
-        Recalculate derived attributes based on multiclass levels.
-        Currently, we update the Base Attack Bonus (BAB) based on the sum of contributions
-        from each class. Using the following simplified progression:
-          - Full: +1 per level.
-          - Average: floor(0.75 * level).
-          - Poor: floor(0.5 * level).
-        The RPGClass data is assumed to be provided by our rpg_class.py configuration.
-        """
         total_bab = 0
-        # For each class, determine the contribution.
         for class_name, level in self.class_levels.items():
             from rpg_class import create_rpg_class
             rpg_class = create_rpg_class(class_name)
@@ -121,14 +111,10 @@ class Character:
             elif progression == "poor":
                 total_bab += int(0.5 * level)
             else:
-                total_bab += level  # Default to full progression.
+                total_bab += level
         self.BAB = total_bab
 
     def level_up(self, rpg_class: "RPGClass") -> None:
-        """
-        Increases the character's level in the given RPG class.
-        Supports multiclassing.
-        """
         class_name = rpg_class.name.lower()
         if class_name in self.class_levels:
             self.class_levels[class_name] += 1
@@ -141,8 +127,9 @@ class Character:
         """
         Computes the character's Armor Class (AC).
         Base AC = 10 + armor_bonus + shield_bonus + natural_armor + deflection_bonus + size_modifier.
-        Adds Dexterity mod and dodge_bonus only if the character is not affected by conditions
-        that remove them (e.g., blinded, flatfooted, paralyzed, unconscious).
+        Adds Dexterity mod and dodge_bonus only if conditions that remove them are absent.
+        For Pathfinder 1e, conditions such as blinded, flatfooted, paralyzed, and unconscious remove the Dex bonus.
+        Stunned does NOT remove the Dex bonus.
         Then adds any additional modifiers from active conditions.
         """
         base_ac = 10 + self.armor_bonus + self.shield_bonus + self.natural_armor + self.deflection_bonus + self.size_modifier
@@ -153,19 +140,12 @@ class Character:
         return base_ac
 
     def get_flatfooted_ac(self) -> int:
-        """
-        Flat-footed AC excludes the Dexterity and dodge bonuses.
-        """
         base_ac = 10 + self.armor_bonus + self.shield_bonus + self.natural_armor + self.deflection_bonus + self.size_modifier
         for cond in self.conditions:
             base_ac += cond.get_modifiers(self).get("ac", 0)
         return base_ac
 
     def get_touch_ac(self) -> int:
-        """
-        Touch AC normally = 10 + Dexterity mod + dodge bonus + size_modifier.
-        If affected by conditions (blinded, flatfooted, paralyzed, unconscious), these bonuses are omitted.
-        """
         base_ac = 10 + self.size_modifier
         if not self.has_condition(["blinded", "flatfooted", "paralyzed", "unconscious"]):
             base_ac += self.get_modifier("DEX") + self.dodge_bonus
